+++ resources/admin.rb
@@ -1,28 +1,13 @@
 # resources/admin.rb
-#
-# Author: Simple Finance <ops@simple.com>
-# License: Apache License, Version 2.0
-#
-# Copyright 2013 Simple Finance Technology Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
+
 # Resource for InfluxDB cluster admin
 
 property :username, String, name_property: true
 property :password, String
 property :auth_username, String, default: 'root'
 property :auth_password, String, default: 'root'
+
+default_action :create
 
 action :create do
   unless password