+++ test/fixtures/cookbooks/influxdb-test/recipes/config_change.rb
@@ -1,0 +1,13 @@
+node.default['influxdb']['config']['graphite'] = [
+  {
+    'enabled' => true
+  }
+]
+influxdb_config node['influxdb']['config_file_path'] do
+  config node['influxdb']['config']
+  notifies :restart, 'service[influxdb]'
+end
+
+service 'influxdb' do
+  action :nothing
+end